@@ -2,64 +2,64 @@
   "name": "crane-ui-plugin",
   "version": "0.0.0",
   "private": true,
-  "repository": "git@github.com:rawagner/crane-ui-plugin.git",
+  "repository": "git@github.com:konveyor/crane-ui-plugin.git",
   "scripts": {
     "clean": "rm -rf ./dist",
-    "build": "yarn clean && NODE_ENV=production yarn ts-node ./node_modules/.bin/webpack",
-    "build-dev": "yarn clean && yarn ts-node ./node_modules/.bin/webpack",
+    "build": "yarn clean && NODE_ENV=production yarn ts-node node_modules/.bin/webpack",
+    "build-dev": "yarn clean && yarn ts-node node_modules/.bin/webpack",
     "http-server": "./http-server.sh ./dist",
-    "ts-node": "ts-node -O '{\"module\":\"commonjs\"}' -I '/node_modules/(?!(@console)/)/'",
+    "ts-node": "./node_modules/.bin/ts-node -O '{\"module\":\"commonjs\"}'",
     "lint": "./node_modules/.bin/eslint .",
     "lint:fix": "yarn lint --fix",
     "format": "./node_modules/.bin/prettier --check --write ./**/*.{tsx,ts,js}",
     "tsc": "./node_modules/.bin/tsc --project ./tsconfig.json --noEmit",
-    "type-check": "yarn tsc",
-    "ci": "yarn type-check && yarn lint"
+    "ci": "yarn tsc && yarn lint"
   },
   "dependencies": {
     "@konveyor/lib-ui": "^5.2.0",
-    "@patternfly/react-core": "^4.162.3",
-    "@patternfly/react-icons": "^4.11.17",
-    "@patternfly/react-table": "^4.31.8",
-    "@patternfly/react-tokens": "^4.12.18",
-    "@types/node": "^14.14.34",
+    "@patternfly/react-core": "^4.175.4",
+    "@patternfly/react-icons": "^4.26.4",
+    "@patternfly/react-table": "^4.44.4",
+    "@patternfly/react-tokens": "^4.27.4",
     "axios": "^0.21.4",
-    "css-loader": "^5.0.1",
     "express": "^4.17.1",
-    "file-loader": "^6.2.0",
     "http-proxy-middleware": "^2.0.1",
     "https-proxy-agent": "^5.0.0",
-    "react": "16.9.0",
+    "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-i18next": "^11.8.11",
     "react-query": "^3.24.4",
     "react-router": "5.1.2",
     "react-router-dom": "5.1.2",
-    "simple-oauth2": "^4.2.0",
-    "style-loader": "^3.3.0",
-    "url-loader": "^4.1.1"
+    "simple-oauth2": "^4.2.0"
   },
   "devDependencies": {
-    "@openshift-console/dynamic-plugin-sdk": "0.0.0-beta1",
+    "@openshift-console/dynamic-plugin-sdk": "0.0.3",
+    "@openshift-console/dynamic-plugin-sdk-webpack": "0.0.4",
+    "@types/node": "^16.11.9",
     "@types/react": "^17.0.30",
     "@types/react-router-dom": "5.1.2",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
     "comment-json": "4.x",
+    "css-loader": "^5.0.1",
     "eslint": "^8.2.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-react-hooks": "^4.2.1-alpha-327d5c484-20211106",
+    "file-loader": "^6.2.0",
     "http-server": "0.12.x",
     "prettier": "^2.4.1",
     "read-pkg": "5.x",
-    "ts-loader": "6.2.2",
+    "style-loader": "^3.3.0",
+    "ts-loader": "^9.2.6",
     "ts-node": "5.0.1",
-    "typescript": "3.8.3",
-    "webpack": "5.0.0-beta.16",
-    "webpack-cli": "4.5.x"
+    "tslib": "^2.3.1",
+    "typescript": "^4.5.2",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.62.1",
+    "webpack-cli": "^4.9.1"
   },
   "consolePlugin": {
     "name": "crane-ui-plugin",
